move lib to src - updated paths
