+++ .npmignore
@@ -1,3 +1,4 @@
 /test
+/src
 .eslintrc.js
 .gitignore
